V2: Corrects wrong SPI mode selection
In the original V1 firmware, SPI mode 0 is used, which assumes that the DAC captures the DIN data on the rising edge of the SCLK clock. However this is not the case. According to the DAC datasheet, the data on the DIN line is captured on the falling edge of the clock. Moreover, in idle mode, the clock line should be at logic high, which is not the case in mode 0. On the other hand, mode 2 meets the DAC requirements. The corrected V2 firmware uses SPI mode 2.