V2: Corrects SPI transmission mode with DAC
In the original V1firmware, SPI transmission mode 0 was used, which assumed that the DAC captures the DIN data on the rising edge of the SCLK clock. However according to the DAC datasheet, the data on the DIN line is captured on the falling edge of the clock. Moreover, in idle mode, the clock line should be at logic high, which is not the case in mode 0. On the other hand, mode 2 meets the DAC requirements. Consequently, mode 2 is now used in V2 firmware.